fix: casts int sur SceneHandle.GetRawData (ulong) + destroy inconditionnel des conditions clonees
GetRawData() renvoie ulong; FishNet stocke ses handles en int -> cast
explicite uniforme. NetworkObserver: le test de signe d instance id
est remplace par le fait que la liste ne contient que des copies
Instantiate() une fois initialisee.

+++ game/Assets/FishNet/Runtime/Managing/Scened/SceneLookupData.cs
@@ -89,7 +89,7 @@ namespace FishNet.Managing.Scened
         /// <param name = "scene">Scene to generate from.</param>
         public SceneLookupData(Scene scene)
         {
-            Handle = scene.handle.GetRawData();
+            Handle = (int)scene.handle.GetRawData();
             Name = scene.name;
         }
